requirements.txt
-i https://pypi.org/simple
apispec==3.3.1
attrs==19.3.0
aws-cdk-aws-apigateway==1.31.0
aws-cdk-aws-batch==1.31.0
aws-cdk-aws-dynamodb==1.31.0
aws-cdk-aws-lambda-event-sources==1.31.0
aws-cdk.assets==1.31.0
aws-cdk.aws-apigateway==1.31.0
aws-cdk.aws-applicationautoscaling==1.31.0
aws-cdk.aws-autoscaling-common==1.31.0
aws-cdk.aws-autoscaling-hooktargets==1.31.0
aws-cdk.aws-autoscaling==1.31.0
aws-cdk.aws-batch==1.31.0
aws-cdk.aws-certificatemanager==1.31.0
aws-cdk.aws-cloudformation==1.31.0
aws-cdk.aws-cloudfront==1.31.0
aws-cdk.aws-cloudwatch==1.31.0
aws-cdk.aws-dynamodb==1.31.0
aws-cdk.aws-ec2==1.31.0
aws-cdk.aws-ecr-assets==1.31.0
aws-cdk.aws-ecr==1.31.0
aws-cdk.aws-ecs==1.31.0
aws-cdk.aws-elasticloadbalancing==1.31.0
aws-cdk.aws-elasticloadbalancingv2==1.31.0
aws-cdk.aws-events==1.31.0
aws-cdk.aws-glue==1.31.0
aws-cdk.aws-iam==1.31.0
aws-cdk.aws-kinesis==1.31.0
aws-cdk.aws-kms==1.31.0
aws-cdk.aws-lambda==1.31.0
aws-cdk.aws-logs==1.31.0
aws-cdk.aws-route53-targets==1.31.0
aws-cdk.aws-route53==1.31.0
aws-cdk.aws-s3-assets==1.31.0
aws-cdk.aws-s3-notifications==1.31.0
aws-cdk.aws-s3==1.31.0
aws-cdk.aws-sam==1.31.0
aws-cdk.aws-secretsmanager==1.31.0
aws-cdk.aws-servicediscovery==1.31.0
aws-cdk.aws-sns-subscriptions==1.31.0
aws-cdk.aws-sns==1.31.0
aws-cdk.aws-sqs==1.31.0
aws-cdk.aws-ssm==1.31.0
aws-cdk.aws-stepfunctions-tasks==1.31.0
aws-cdk.aws-stepfunctions==1.31.0
aws-cdk.core==1.31.0
aws-cdk.custom-resources==1.31.0
aws-cdk.cx-api==1.31.0
aws-cdk.region-info==1.31.0
aws-lambda-wsgi==0.0.6
boto3==1.14.12
botocore==1.17.12
cattrs==1.0.0
certifi==2020.6.20
chardet==3.0.4
click==7.1.1
coloredlogs==14.0
constructs==2.0.0
docutils==0.15.2
flask-apispec==0.9.0
flask-cors==3.0.8
flask-serverless==0.1.4
flask==1.1.2
humanfriendly==8.2
idna==2.10
itsdangerous==1.1.0
jinja2==2.11.2
jmespath==0.10.0
jsii==1.1.0
markupsafe==1.1.1
marshmallow==3.6.1 ; python_version >= '3'
publication==0.0.3
pynamodb==4.3.2
python-dateutil==2.8.1
requests==2.24.0
s3transfer==0.3.3
six==1.14.0
typing-extensions==3.7.4.1
urllib3==1.25.8
webargs==5.5.3
werkzeug==1.0.1
